pyproject.toml
[tool.pytest.ini_options]
addopts = "--cov=src/brainxio --timeout=120"
timeout = 120

[tool.coverage.run]
source = ["src/brainxio"]

[tool.coverage.report]
show_missing = true

[tool.poetry]
name = "brainxio"
version = "0.1.0"
description = "A modular CLI for automation and AI-driven tasks"
authors = ["Your Name <you@example.com>"]
license = "Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3"
pytest-cov = "^5.0"
pytest-timeout = "^2.3"
pytest-reraise = "^2.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
